fix: refactorizado studentProfile en su modulo correspondiente y corregido errores
