+++ backend/apps/common/views.py
@@ -1,6 +1,10 @@
 # apps/common/views.py
+import base64
+import logging
+import uuid
 from django.conf import settings
 from django.contrib.auth import get_user_model
+from django.core.files.base import ContentFile
 from rest_framework import status
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
@@ -9,6 +13,7 @@ from rest_framework.views import APIView
 from apps.common.utils.jwt_auth import resolve_user_from_request
 from apps.membership.models import Membership, Role
 from apps.residences.models import ResidenceBranding
+from apps.residents.models import StudentProfile
 
 from .serializers import (
     AdminCreateResidentSerializer,
@@ -18,6 +23,7 @@ from .serializers import (
     PasswordResetConfirmSerializer,
     PasswordResetRequestSerializer,
     PlanSerializer,
+    StudentProfileSerializer,
 )
 from .services import (
     authenticate_user,
@@ -28,6 +34,7 @@ from .services import (
 )
 
 UserModel = get_user_model()
+logger = logging.getLogger(__name__)
 
 
 class TenantContextView(APIView):
@@ -151,7 +158,7 @@ class AuthLoginView(APIView):
 
         residence = getattr(request, "residence", None)
         if not has_access_for_portal(user, portal, residence):
-            from apps.membership.models import Membership
+            
             has_any_access = False
             fallback_residence = None
 
@@ -376,21 +383,26 @@ class StudentProfileView(APIView):
                 {"detail": "No autenticado."}, status=status.HTTP_401_UNAUTHORIZED
             )
 
-        from apps.residences.models import StudentProfile
+        user_pk = user_data.get("user_id") or user_data.get("id") or user_data.get("sub")
+        try:
+            user = UserModel.objects.get(pk=user_pk)
+        except UserModel.DoesNotExist:
+            return Response(
+                {"detail": "Usuario no encontrado."}, status=status.HTTP_404_NOT_FOUND
+            )
 
         try:
-            user_pk = (
-                user_data.get("user_id") or user_data.get("id") or user_data.get("sub")
+            profile, _ = StudentProfile.objects.get_or_create(
+                user=user,
+                defaults={"residence": getattr(request, "residence", None)},
             )
-            user = UserModel.objects.get(pk=user_pk)
-            profile = StudentProfile.objects.get(user=user)
-            from .serializers import StudentProfileSerializer
-
             serializer = StudentProfileSerializer(profile)
             return Response(serializer.data, status=status.HTTP_200_OK)
-        except StudentProfile.DoesNotExist:
+        except Exception:
+            logger.exception("Error retrieving student profile for user_id=%s", user_pk)
             return Response(
-                {"detail": "Perfil no encontrado."}, status=status.HTTP_404_NOT_FOUND
+                {"detail": "Error al procesar el perfil."},
+                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
 
     def post(self, request):
@@ -401,25 +413,52 @@ class StudentProfileView(APIView):
                 {"detail": "No autenticado."}, status=status.HTTP_401_UNAUTHORIZED
             )
 
-        from apps.residences.models import StudentProfile
-
-        from .serializers import StudentProfileSerializer
-
         user_pk = (
             user_data.get("user_id") or user_data.get("id") or user_data.get("sub")
         )
-        user = UserModel.objects.get(pk=user_pk)
+        try:
+            user = UserModel.objects.get(pk=user_pk)
+        except UserModel.DoesNotExist:
+            return Response(
+                {"detail": "Usuario no encontrado."}, status=status.HTTP_404_NOT_FOUND
+            )
 
         request.user = user
+
+        mutable_data = request.data.copy()
+        profile_image_b64 = mutable_data.pop("profile_image", None)
+
+        # Handle base64 image if present
+        if (
+            profile_image_b64
+            and isinstance(profile_image_b64, str)
+            and profile_image_b64.startswith("data:image")
+        ):
+            try:
+                image_format, imgstr = profile_image_b64.split(";base64,")
+                ext = image_format.split("/")[-1]
+                ext = ext.split(";")[0] if ";" in ext else ext
+                file_name = f"{uuid.uuid4().hex}.{ext}"
+                mutable_data["profile_image"] = ContentFile(
+                    base64.b64decode(imgstr), name=file_name
+                )
+            except Exception:
+                logger.exception(
+                    "Failed to decode student profile image for user_id=%s", user_pk
+                )
+                return Response(
+                    {"profile_image": ["Formato de imagen no valido."]},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
 
         try:
             profile = StudentProfile.objects.get(user=user)
             serializer = StudentProfileSerializer(
-                profile, data=request.data, partial=True, context={"request": request}
+                profile, data=mutable_data, partial=True, context={"request": request}
             )
         except StudentProfile.DoesNotExist:
             serializer = StudentProfileSerializer(
-                data=request.data, context={"request": request}
+                data=mutable_data, context={"request": request}
             )
 
         if serializer.is_valid():
