+++ backend/apps/events/tests.py
@@ -1184,7 +1184,7 @@ class UrlsTests(FastTenantTestCase):
         )
 
 
-from apps.residences.models import StudentProfile
+from apps.residents.models import StudentProfile
 
 
 class EventRecommendationTests(ViewTestsBase):
